sch: move all lcore code to sch
