+++ lib/src/dev/mt_dev.c
@@ -1157,255 +1157,6 @@ static int dev_start_port(struct mt_interface* inf) {
   return 0;
 }
 
-static int dev_filelock_lock(struct mtl_main_impl* impl) {
-  int fd = open(MT_FLOCK_PATH, O_RDONLY | O_CREAT, 0666);
-  if (fd < 0) {
-    /* sometimes may fail due to user permission, try open read-only */
-    fd = open(MT_FLOCK_PATH, O_RDONLY);
-    if (fd < 0) {
-      err("%s, failed to open %s, %s\n", __func__, MT_FLOCK_PATH, strerror(errno));
-      return -EIO;
-    }
-  }
-  impl->lcore_lock_fd = fd;
-  /* wait until locked */
-  if (flock(fd, LOCK_EX) != 0) {
-    err("%s, can not lock file\n", __func__);
-    close(fd);
-    impl->lcore_lock_fd = -1;
-    return -EIO;
-  }
-
-  return 0;
-}
-
-static int dev_filelock_unlock(struct mtl_main_impl* impl) {
-  int fd = impl->lcore_lock_fd;
-
-  if (fd < 0) {
-    err("%s, wrong lock file fd %d\n", __func__, fd);
-    return -EIO;
-  }
-
-  if (flock(fd, LOCK_UN) != 0) {
-    err("%s, can not unlock file\n", __func__);
-    return -EIO;
-  }
-  close(fd);
-  impl->lcore_lock_fd = -1;
-  return 0;
-}
-
-int mt_dev_get_lcore(struct mtl_main_impl* impl, unsigned int* lcore) {
-  unsigned int cur_lcore = 0;
-  int ret;
-  struct mt_lcore_shm* lcore_shm = impl->lcore_shm;
-
-  ret = dev_filelock_lock(impl);
-  if (ret < 0) {
-    err("%s, dev_filelock_lock fail\n", __func__);
-    return ret;
-  }
-
-  do {
-    cur_lcore = rte_get_next_lcore(cur_lcore, 1, 0);
-
-    if ((cur_lcore < RTE_MAX_LCORE) && mt_socket_match(rte_lcore_to_socket_id(cur_lcore),
-                                                       mt_socket_id(impl, MTL_PORT_P))) {
-      if (!lcore_shm->lcores_active[cur_lcore]) {
-        *lcore = cur_lcore;
-        lcore_shm->lcores_active[cur_lcore] = true;
-        lcore_shm->used++;
-        rte_atomic32_inc(&impl->lcore_cnt);
-        impl->local_lcores_active[cur_lcore] = true;
-        ret = dev_filelock_unlock(impl);
-        info("%s, available lcore %d\n", __func__, cur_lcore);
-        if (ret < 0) {
-          err("%s, dev_filelock_unlock fail\n", __func__);
-          return ret;
-        }
-        return 0;
-      }
-    }
-  } while (cur_lcore < RTE_MAX_LCORE);
-
-  dev_filelock_unlock(impl);
-  err("%s, fail to find lcore\n", __func__);
-  return -EIO;
-}
-
-int mt_dev_put_lcore(struct mtl_main_impl* impl, unsigned int lcore) {
-  int ret;
-  struct mt_lcore_shm* lcore_shm = impl->lcore_shm;
-
-  if (lcore >= RTE_MAX_LCORE) {
-    err("%s, invalid lcore %d\n", __func__, lcore);
-    return -EIO;
-  }
-  if (!lcore_shm) {
-    err("%s, no lcore shm attached\n", __func__);
-    return -EIO;
-  }
-  ret = dev_filelock_lock(impl);
-  if (ret < 0) {
-    err("%s, dev_filelock_lock fail\n", __func__);
-    return ret;
-  }
-  if (!lcore_shm->lcores_active[lcore]) {
-    err("%s, lcore %d not active\n", __func__, lcore);
-    ret = -EIO;
-    goto err_unlock;
-  }
-
-  lcore_shm->lcores_active[lcore] = false;
-  lcore_shm->used--;
-  rte_atomic32_dec(&impl->lcore_cnt);
-  impl->local_lcores_active[lcore] = false;
-  ret = dev_filelock_unlock(impl);
-  info("%s, lcore %d\n", __func__, lcore);
-  if (ret < 0) {
-    err("%s, dev_filelock_unlock fail\n", __func__);
-    return ret;
-  }
-  return 0;
-
-err_unlock:
-  dev_filelock_unlock(impl);
-  return ret;
-}
-
-bool mt_dev_lcore_valid(struct mtl_main_impl* impl, unsigned int lcore) {
-  struct mt_lcore_shm* lcore_shm = impl->lcore_shm;
-
-  if (lcore >= RTE_MAX_LCORE) {
-    err("%s, invalid lcore %d\n", __func__, lcore);
-    return -EIO;
-  }
-  if (!lcore_shm) {
-    err("%s, no lcore shm attached\n", __func__);
-    return -EIO;
-  }
-
-  return lcore_shm->lcores_active[lcore];
-}
-
-static int dev_uinit_lcores(struct mtl_main_impl* impl) {
-  int ret;
-  int shm_id = impl->lcore_shm_id;
-  struct mt_lcore_shm* lcore_shm = impl->lcore_shm;
-  if (!lcore_shm || shm_id == -1) {
-    err("%s, no lcore shm attached\n", __func__);
-    return -EIO;
-  }
-
-  for (unsigned int lcore = 0; lcore < RTE_MAX_LCORE; lcore++) {
-    if (impl->local_lcores_active[lcore]) {
-      warn("%s, lcore %d still active\n", __func__, lcore);
-      mt_dev_put_lcore(impl, lcore);
-    }
-  }
-
-  ret = dev_filelock_lock(impl);
-  if (ret < 0) {
-    err("%s, dev_filelock_lock fail\n", __func__);
-    return ret;
-  }
-
-  ret = shmdt(impl->lcore_shm);
-  if (ret == -1) {
-    err("%s, shared memory detach failed, %s\n", __func__, strerror(errno));
-    goto err_unlock;
-  }
-
-  struct shmid_ds shmds;
-  ret = shmctl(shm_id, IPC_STAT, &shmds);
-  if (ret < 0) {
-    err("%s, can not stat shared memory, %s\n", __func__, strerror(errno));
-    goto err_unlock;
-  }
-  if (shmds.shm_nattch == 0) { /* remove ipc if we are the last user */
-    ret = shmctl(shm_id, IPC_RMID, NULL);
-    if (ret < 0) {
-      warn("%s, can not remove shared memory, %s\n", __func__, strerror(errno));
-      goto err_unlock;
-    }
-  }
-
-  impl->lcore_shm_id = -1;
-  impl->lcore_shm = NULL;
-  ret = dev_filelock_unlock(impl);
-  if (ret < 0) {
-    err("%s, dev_filelock_unlock fail\n", __func__);
-    return ret;
-  }
-  return 0;
-
-err_unlock:
-  dev_filelock_unlock(impl);
-  return ret;
-}
-
-static int dev_init_lcores(struct mtl_main_impl* impl) {
-  int ret;
-  struct mt_lcore_shm* lcore_shm;
-
-  if (impl->lcore_shm) {
-    err("%s, lcore_shm attached\n", __func__);
-    return -EIO;
-  }
-
-  ret = dev_filelock_lock(impl);
-  if (ret < 0) {
-    err("%s, dev_filelock_lock fail\n", __func__);
-    return ret;
-  }
-
-  key_t key = ftok("/dev/null", 21);
-  if (key < 0) {
-    err("%s, ftok error: %s\n", __func__, strerror(errno));
-    ret = -EIO;
-    goto err_unlock;
-  }
-  int shm_id = shmget(key, sizeof(*lcore_shm), 0666 | IPC_CREAT);
-  if (shm_id < 0) {
-    err("%s, can not get shared memory for lcore, %s\n", __func__, strerror(errno));
-    ret = -EIO;
-    goto err_unlock;
-  }
-  impl->lcore_shm_id = shm_id;
-
-  lcore_shm = shmat(shm_id, NULL, 0);
-  if (lcore_shm == (void*)-1) {
-    err("%s, can not attach shared memory for lcore, %s\n", __func__, strerror(errno));
-    ret = -EIO;
-    goto err_unlock;
-  }
-
-  struct shmid_ds stat;
-  ret = shmctl(shm_id, IPC_STAT, &stat);
-  if (ret < 0) {
-    err("%s, shmctl fail\n", __func__);
-    shmdt(lcore_shm);
-    goto err_unlock;
-  }
-  if (stat.shm_nattch == 1) /* clear shm as we are the first user */
-    memset(lcore_shm, 0, sizeof(*lcore_shm));
-
-  impl->lcore_shm = lcore_shm;
-  info("%s, shared memory attached at %p nattch %d\n", __func__, impl->lcore_shm,
-       (int)stat.shm_nattch);
-  ret = dev_filelock_unlock(impl);
-  if (ret < 0) {
-    err("%s, dev_filelock_unlock fail\n", __func__);
-    return ret;
-  }
-  return 0;
-
-err_unlock:
-  dev_filelock_unlock(impl);
-  return ret;
-}
-
 static int dev_if_uinit_rx_queues(struct mt_interface* inf) {
   enum mtl_port port = inf->port;
   struct mt_rx_queue* rx_queue;
@@ -1991,9 +1742,6 @@ int mt_dev_create(struct mtl_main_impl* impl) {
   struct mt_interface* inf;
   enum mt_port_type port_type;
 
-  ret = dev_init_lcores(impl);
-  if (ret < 0) return ret;
-
   for (int i = 0; i < num_ports; i++) {
     int detect_retry = 0;
 
@@ -2112,7 +1860,6 @@ int mt_dev_free(struct mtl_main_impl* impl) {
   struct mt_interface* inf;
 
   mt_sch_mrg_uinit(impl);
-  dev_uinit_lcores(impl);
 
   for (int i = 0; i < num_ports; i++) {
     inf = mt_if(impl, i);
