+++ lib/src/st2110/st_rx_video_session.c
@@ -2437,7 +2437,7 @@ static int rv_uinit_pkt_lcore(struct mtl_main_impl* impl,
 
   if (s->has_pkt_lcore) {
     rte_eal_wait_lcore(s->pkt_lcore);
-    mt_dev_put_lcore(impl, s->pkt_lcore);
+    mt_sch_put_lcore(impl, s->pkt_lcore);
     s->has_pkt_lcore = false;
   }
 
@@ -2488,7 +2488,7 @@ static int rv_init_pkt_lcore(struct mtl_main_impl* impl,
   }
   s->pkt_lcore_ring = ring;
 
-  ret = mt_dev_get_lcore(impl, &lcore);
+  ret = mt_sch_get_lcore(impl, &lcore);
   if (ret < 0) {
     err("%s(%d,%d), get lcore fail %d\n", __func__, mgr_idx, idx, ret);
     rv_uinit_pkt_lcore(impl, s);
@@ -3043,7 +3043,7 @@ static int rv_init_rtcp_uhdr(struct mtl_main_impl* impl,
   struct rte_ether_addr* d_addr = mt_eth_d_addr(eth);
 
   /* ether hdr */
-  ret = mt_dst_ip_mac(impl, dip, d_addr, port, MT_DEV_TIMEOUT_INFINITE);
+  ret = mt_dst_ip_mac(impl, dip, d_addr, port, MT_TIMEOUT_INFINITE);
   if (ret < 0) {
     err("%s(%d), get mac fail %d for %d.%d.%d.%d\n", __func__, idx, ret, dip[0], dip[1],
         dip[2], dip[3]);
